Moving config for Status Leds
